# Faster `linkerd install --ignore-cluster`