The `linkerd install` `--ignore-cluster` and `--skip-checks` flags
enable generating install manifests without a connection to a k8s
cluster. Unfortunately these flags were only checked after attempted
connections to a k8s cluster were made. This satisfied the use case of
`linkerd install` "ignoring" the state of the cluster, but for
environments not connected to a cluster, the user would have to wait for
30s timeouts before getting the manifests.

Modify `linkerd install` and its subcommands to pre-emptively check for
`--ignore-cluster` and `--skip-checks`. This decreases `linkerd install
--ignore-cluster` from ~30s to ~1s, and `linkerd install control-plane
--ignore-cluster --skip-checks` from ~60s to ~1s.

Signed-off-by: Andrew Seigner <siggy@buoyant.io>

## before

```
$ time bin/go-run cli install --ignore-cluster > /dev/null

real	0m35.948s
user	0m5.518s
sys	0m2.085s

$ time bin/go-run cli install control-plane --ignore-cluster --skip-checks > /dev/null

real	1m3.408s
user	0m2.723s
sys	0m1.979s
```

## after

```
$ time bin/go-run cli install --ignore-cluster > /dev/null

real	0m4.140s
user	0m3.887s
sys	0m2.027s

$ time bin/go-run cli install control-plane --ignore-cluster --skip-checks > /dev/null

real	0m3.237s
user	0m2.631s
sys	0m1.938s
```